fixes to models

+++ twitter_registration/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 
 class TwitterUser(models.Model):
 	user = models.OneToOneField(User, related_name='twitter_user')
-	twitter_id = models.IntegerField()
+	twitter_id = models.BigIntegerField()
 	access_token = models.CharField(max_length=200)
 	access_token_secret = models.CharField(max_length=200)
